requirements.txt
Django==3.0
django-redis==4.11.0
django-rest-swagger==2.2.0
django-split-settings==0.3.0
djangorestframework==3.12.4
psycopg2-binary==2.8.6
psycopg2==2.8.6
requests==2.19.1
requests[security]
pillow

# Testing
coverage
django-debug-toolbar
pytest==3.0.7
pytest-django==3.1.2
mixer==6.1.3
pytest-mock==1.10.0
pylint==1.8.4
pylint-django==0.9.4